database: add support for PostgreSQL backend
This commit introduces the ability to use a PostgreSQL backend. It
abstracts all of the database operations so that in the future we could
theoretically add additional backends as well.

The PostgreSQL backend has some notable advantages over the SQLite
backend including using the UUID type in the database, and using the
TIMESTAMP type in most cases as well.

Co-authored-by: finn <finn@beeper.com>